backend/requirements.txt
# Django
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
django-filter>=24.0,<25.0
djangorestframework-simplejwt>=5.3,<6.0

# Channels (WebSocket)
daphne>=4.1,<5.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0

# Billing
dj-stripe>=2.9,<3.0

# Database
psycopg[binary]>=3.2,<4.0
dj-database-url>=2.2,<3.0

# Environment
python-dotenv>=1.0,<2.0

# Server
gunicorn>=22.0,<23.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-asyncio>=0.24,<1.0
factory-boy>=3.3,<4.0

# Utilities
django-extensions>=3.2,<4.0
